add html template for forms

+++ happyrabbit/hr_user/models.py
@@ -8,7 +8,7 @@ from .enums import EXTERNAL_SERVICE_CHOICES
 
 class Account(ExternalAccount, models.Model):
 
-    account_id = models.IntegerField(primary_key=True)
+    account_id = models.BigAutoField(primary_key=True)
     external_service = models.CharField(max_length=200, choices=EXTERNAL_SERVICE_CHOICES)
     external_user_id = models.IntegerField(blank=True)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
@@ -29,7 +29,7 @@ class Account(ExternalAccount, models.Model):
 
 class UserProfile(models.Model, ExternalUserProfile):
 
-    user_profile_id = models.IntegerField(primary_key=True)
+    user_profile_id = models.BigAutoField(primary_key=True)
     first_name = models.CharField(max_length=200, blank=True)
     last_name = models.CharField(max_length=200, blank=True)
     language_code = models.CharField(max_length=2, blank=True)
@@ -50,7 +50,7 @@ class UserProfile(models.Model, ExternalUserProfile):
 
 
 class Child(models.Model):
-    child_id = models.IntegerField(primary_key=True)
+    child_id = models.BigAutoField(primary_key=True)
     name = models.CharField(max_length=200)
     age = models.IntegerField()
     carrots = models.IntegerField(default=0)
